+++ src/Symfony/Bridge/Doctrine/ArgumentResolver/EntityValueResolver.php
@@ -17,11 +17,13 @@ use Doctrine\ORM\NoResultException;
 use Doctrine\Persistence\ManagerRegistry;
 use Doctrine\Persistence\ObjectManager;
 use Symfony\Bridge\Doctrine\Attribute\MapEntity;
+use Symfony\Component\DependencyInjection\Argument\ServiceLocator;
 use Symfony\Component\ExpressionLanguage\ExpressionLanguage;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpKernel\Controller\ValueResolverInterface;
 use Symfony\Component\HttpKernel\ControllerMetadata\ArgumentMetadata;
 use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
+
 
 /**
  * Yields the entity matching the criteria provided in the route.
@@ -35,7 +37,8 @@ final class EntityValueResolver implements ValueResolverInterface
         private ManagerRegistry $registry,
         private ?ExpressionLanguage $expressionLanguage = null,
         private MapEntity $defaults = new MapEntity(),
-        private ?EntityValueResolverVariableInjectorInterface $variableInjector = null,
+        private ?ServiceLocator $variablesInjectorLocator = null,
+        private array $defaulInjectors = [],
     ) {
     }
 
@@ -58,9 +61,9 @@ final class EntityValueResolver implements ValueResolverInterface
         $message = '';
         if (null !== $options->expr) {
             if (null === $object = $this->findViaExpression($manager, $request, $options)) {
-                $message = \sprintf(' The expression "%s" returned null.', $options->expr);
+                $message = sprintf(' The expression "%s" returned null.', $options->expr);
             }
-        // find by identifier?
+            // find by identifier?
         } elseif (false === $object = $this->find($manager, $request, $options, $argument)) {
             // find by criteria
             if (!$criteria = $this->getCriteria($request, $options, $manager, $argument)) {
@@ -74,7 +77,7 @@ final class EntityValueResolver implements ValueResolverInterface
         }
 
         if (null === $object && !$argument->isNullable()) {
-            throw new NotFoundHttpException($options->message ?? (\sprintf('"%s" object not found by "%s".', $options->class, self::class).$message));
+            throw new NotFoundHttpException($options->message ?? (sprintf('"%s" object not found by "%s".', $options->class, self::class).$message));
         }
 
         return [$object];
@@ -127,7 +130,7 @@ final class EntityValueResolver implements ValueResolverInterface
             foreach ($options->id as $field) {
                 // Convert "%s_uuid" to "foobar_uuid"
                 if (str_contains($field, '%s')) {
-                    $field = \sprintf($field, $argument->getName());
+                    $field = sprintf($field, $argument->getName());
                 }
 
                 $id[$field] = $request->attributes->get($field);
@@ -215,11 +218,25 @@ final class EntityValueResolver implements ValueResolverInterface
     private function findViaExpression(ObjectManager $manager, Request $request, MapEntity $options): object|iterable|null
     {
         if (!$this->expressionLanguage) {
-            throw new \LogicException(\sprintf('You cannot use the "%s" if the ExpressionLanguage component is not available. Try running "composer require symfony/expression-language".', __CLASS__));
+            throw new \LogicException(sprintf('You cannot use the "%s" if the ExpressionLanguage component is not available. Try running "composer require symfony/expression-language".', __CLASS__));
+        }
+
+        $variablesFormInjectors = [];
+        foreach($options->modifiersInjectors ?? $this->defaulInjectors as $injector) {
+            $injectorObject = $this->variablesInjectorLocator->get($injector);
+
+            if (!$injectorObject instanceof EntityValueResolverExpressionModifiersInjectorInterface) {
+                throw new \LogicException(sprintf('You cannot use "%s" as Expression Modifier, it must implement %s.', $injector, EntityValueResolverExpressionModifiersInjectorInterface::class));
+            }
+
+            $variablesFormInjectors = array_merge($variablesFormInjectors, $injectorObject->getVariables());
+            foreach($injectorObject->getFunctions() as $injectorFunction) {
+                $this->expressionLanguage->addFunction($injectorFunction);
+            }
         }
 
         $repository = $manager->getRepository($options->class);
-        $variables = array_merge($this->variableInjector?->getVariables() ?? [], $request->attributes->all(), [
+        $variables = array_merge($variablesFormInjectors, $request->attributes->all(), [
             'repository' => $repository,
             'request' => $request,
         ]);